sentiment-service/requirements.txt
fastapi
uvicorn[standard]
pydantic
pymongo[srv]
dnspython
requests
python-dotenv
huggingface-hub